measurements: Refactor getting matching filters into separate function
Doing this in preparation for using the same functions to get the
filter measurements within `applyMeasurementsColorBy` in the following
commit.
